+++ src/main/java/net/oneandone/lavender/filter/processor/UriHelper.java
@@ -42,4 +42,6 @@ public class UriHelper {
         return resolved;
     }
 
+    private UriHelper() {
+    }
 }