+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 2.7.2 (pending)
 
-* update lazy-foss-parent 1.0.4 to 1.1.0
+* update lazy-foss-parent 1.0.4 to 1.2.0
 
 
 ### 2.7.1 (2018-11-22)