+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * fixed NumberFormatException when creating released git modules
 * update lazy-foss-parent 1.0.4 to 1.2.0
+* checkstyle fixes
 
 
 ### 2.7.1 (2018-11-22)